Replaced GiftCertificateUpdateParametersHolder with GiftCertificate in update()

+++ gift-certificate-data/src/main/java/com/epam/esm/util/DAOUtils.java
@@ -27,7 +27,7 @@ public final class DAOUtils {
         List<GiftCertificate> certificates = new ArrayList<>();
         while (resultSet.next()) {
             GiftCertificate tempCertificate = new GiftCertificate(
-                    resultSet.getInt("id"),
+                    resultSet.getLong("id"),
                     resultSet.getString("name"),
                     resultSet.getString("description"),
                     resultSet.getDouble("price"),
